pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.5.0",
    "psycopg2-binary>=2.9.11",
    "pydub>=0.25.1",
    "reportlab>=4.4.4",
    "sqlalchemy>=2.0.44",
    "streamlit-audiorec>=0.1.3",
    "streamlit>=1.50.0",
    "pytesseract>=0.3.13",
    "pillow>=11.3.0",
]
